So ALOT changed lol
I decided to change to algorithm to be: Find the max value in the column. compare that max value to the element to its right. if its greater...continue with the previous algorithm. If it is not...get the max value again from a pool of without the previous max value